opt/airflow/requirements.txt
anyio==4.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
contourpy==1.1.1
cycler==0.12.1
fonttools==4.53.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
kiwisolver==1.4.5
matplotlib==3.4.3
numpy==1.24.4
packaging==24.1
pandas==1.5.3
pillow==10.4.0
psycopg2-binary==2.9.9
pyarrow==10.0.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==13.7
pytz==2024.1
requests==2.32.3
seaborn==0.13.2
six==1.16.0
sniffio==1.3.1
tzdata==2024.1
urllib3==2.2.2